+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/ApiServices.scala
@@ -107,7 +107,7 @@ private[daml] object ApiServices {
 
       for {
         ledgerId <- Resource.fromFuture(indexService.getLedgerId())
-        ledgerConfigProvider <- LedgerConfigProvider
+        currentLedgerConfiguration <- LedgerConfigProvider
           .owner(
             indexService,
             optWriteService,
@@ -116,7 +116,7 @@ private[daml] object ApiServices {
           )(materializer, loggingContext)
           .acquire()
         services <- Resource(
-          Future(createServices(ledgerId, ledgerConfigProvider)(servicesExecutionContext))
+          Future(createServices(ledgerId, currentLedgerConfiguration)(servicesExecutionContext))
         )(services =>
           Future {
             services.foreach {
@@ -130,7 +130,7 @@ private[daml] object ApiServices {
 
     private def createServices(
         ledgerId: LedgerId,
-        ledgerConfigProvider: LedgerConfigProvider,
+        currentLedgerConfiguration: CurrentLedgerConfiguration,
     )(implicit executionContext: ExecutionContext): List[BindableService] = {
       val apiTransactionService =
         ApiTransactionService.create(ledgerId, transactionsService, metrics)
@@ -160,7 +160,7 @@ private[daml] object ApiServices {
       val writeServiceBackedApiServices =
         intitializeWriteServiceBackedApiServices(
           ledgerId,
-          ledgerConfigProvider,
+          currentLedgerConfiguration,
           apiCompletionService,
           apiTransactionService,
         )
@@ -186,7 +186,7 @@ private[daml] object ApiServices {
 
     private def intitializeWriteServiceBackedApiServices(
         ledgerId: LedgerId,
-        ledgerConfigProvider: LedgerConfigProvider,
+        currentLedgerConfiguration: CurrentLedgerConfiguration,
         apiCompletionService: GrpcCommandCompletionService,
         apiTransactionService: GrpcTransactionService,
     )(implicit executionContext: ExecutionContext): List[BindableService] = {
@@ -214,7 +214,7 @@ private[daml] object ApiServices {
           partyManagementService,
           timeProvider,
           timeProviderType,
-          ledgerConfigProvider,
+          currentLedgerConfiguration,
           seedService,
           commandExecutor,
           ApiSubmissionService.Configuration(
@@ -243,7 +243,7 @@ private[daml] object ApiServices {
             apiTransactionService.getFlatTransactionById,
           ),
           timeProvider,
-          ledgerConfigProvider,
+          currentLedgerConfiguration,
           metrics,
         )
