+++ ledger/participant-integration-api/src/main/scala/platform/apiserver/services/ApiCommandService.scala
@@ -50,7 +50,7 @@ import scala.util.Try
 private[apiserver] final class ApiCommandService private (
     services: LocalServices,
     configuration: ApiCommandService.Configuration,
-    ledgerConfigProvider: LedgerConfigProvider,
+    currentLedgerConfiguration: CurrentLedgerConfiguration,
     metrics: Metrics,
 )(implicit
     materializer: Materializer,
@@ -86,7 +86,7 @@ private[apiserver] final class ApiCommandService private (
       logging.readAsStrings(request.getCommands.readAs),
     ) { implicit loggingContext =>
       if (running) {
-        ledgerConfigProvider.latestConfiguration.fold[Future[Completion]](
+        currentLedgerConfiguration.latestConfiguration.fold[Future[Completion]](
           Future.failed(ErrorFactories.missingLedgerConfig())
         )(ledgerConfig => track(request, ledgerConfig))
       } else {
@@ -196,7 +196,7 @@ private[apiserver] object ApiCommandService {
       configuration: Configuration,
       services: LocalServices,
       timeProvider: TimeProvider,
-      ledgerConfigProvider: LedgerConfigProvider,
+      currentLedgerConfiguration: CurrentLedgerConfiguration,
       metrics: Metrics,
   )(implicit
       materializer: Materializer,
@@ -204,12 +204,12 @@ private[apiserver] object ApiCommandService {
       loggingContext: LoggingContext,
   ): CommandServiceGrpc.CommandService with GrpcApiService =
     new GrpcCommandService(
-      new ApiCommandService(services, configuration, ledgerConfigProvider, metrics),
+      new ApiCommandService(services, configuration, currentLedgerConfiguration, metrics),
       ledgerId = configuration.ledgerId,
       currentLedgerTime = () => timeProvider.getCurrentTime,
       currentUtcTime = () => Instant.now,
       maxDeduplicationTime = () =>
-        ledgerConfigProvider.latestConfiguration.map(_.maxDeduplicationTime),
+        currentLedgerConfiguration.latestConfiguration.map(_.maxDeduplicationTime),
       generateSubmissionId = SubmissionIdGenerator.Random,
     )
